zcash_protocol::consensus

Enum BranchId

pub enum BranchId {
    Sprout,
    Overwinter,
    Sapling,
    Blossom,
    Heartwood,
    Canopy,
    Nu5,
    Nu6,
}

A globally-unique identifier for a set of consensus rules within the Zcash chain.

Each branch ID in this enum corresponds to one of the epochs between a pair of Zcash network upgrades. For example, BranchId::Overwinter corresponds to the blocks starting at Overwinter activation, and ending the block before Sapling activation.

The main use of the branch ID is in signature generation: transactions commit to a specific branch ID by including it as part of signature_hash. This ensures two-way replay protection for transactions across network upgrades.

See ZIP 200 for more details.

Variants

Sprout

The consensus rules at the launch of Zcash.

Overwinter

The consensus rules deployed by NetworkUpgrade::Overwinter.

Sapling

The consensus rules deployed by NetworkUpgrade::Sapling.

Blossom

The consensus rules deployed by NetworkUpgrade::Blossom.

Heartwood

The consensus rules deployed by NetworkUpgrade::Heartwood.

Canopy

The consensus rules deployed by NetworkUpgrade::Canopy.

Nu5

The consensus rules deployed by NetworkUpgrade::Nu5.

Nu6

The consensus rules deployed by NetworkUpgrade::Nu6.

Implementations

impl BranchId

pub fn for_height<P: Parameters>(parameters: &P, height: BlockHeight) -> Self

Returns the branch ID corresponding to the consensus rule set that is active at the given height.

This is the branch ID that should be used when creating transactions.

pub fn height_range<P: Parameters>( &self, params: &P, ) -> Option<impl RangeBounds<BlockHeight>>

Returns the range of heights for the consensus epoch associated with this branch id.

The resulting tuple implements the RangeBounds<BlockHeight> trait.

pub fn height_bounds<P: Parameters>( &self, params: &P, ) -> Option<(BlockHeight, Option<BlockHeight>)>

Returns the range of heights for the consensus epoch associated with this branch id.

The return type of this value is slightly more precise than Self::height_range:

Some((x, Some(y))) means that the consensus rules corresponding to this branch id are in effect for the range x..y

Some((x, None)) means that the consensus rules corresponding to this branch id are in effect for the range x..

None means that the consensus rules corresponding to this branch id are never in effect.

pub fn sprout_uses_groth_proofs(&self) -> bool