zcash_protocol
Module value

Modules

testingtest-dependencies

Structs

QuotRem

A struct that provides both the quotient and remainder of a division operation.

ZatBalance

A type-safe representation of a Zcash value delta, in zatoshis.

Zatoshis

A type-safe representation of some nonnegative amount of Zcash.

Enums

BalanceError

A type for balance violations in amount addition and subtraction (overflow and underflow of allowed ranges)

Constants

COIN

MAX_BALANCE

MAX_MONEY