zcash_protocol::value

Struct Zatoshis

pub struct Zatoshis(/* private fields */);

A type-safe representation of some nonnegative amount of Zcash.

A Zatoshis can only be constructed from an integer that is within the valid monetary range of {0..MAX_MONEY} (where MAX_MONEY = 21,000,000 × 10⁸ zatoshis).

Implementations

impl Zatoshis

pub const ZERO: Self

Returns the identity Zatoshis

pub fn into_u64(self) -> u64

Returns this Zatoshis as a u64.

pub fn from_u64(amount: u64) -> Result<Self, BalanceError>

Creates a Zatoshis from a u64.

Returns an error if the amount is outside the range {0..MAX_MONEY}.

pub const fn const_from_u64(amount: u64) -> Self

Creates a constant Zatoshis from a u64.

Panics: if the amount is outside the range {0..MAX_MONEY}.

pub fn from_nonnegative_i64(amount: i64) -> Result<Self, BalanceError>

Creates a Zatoshis from an i64.

Returns an error if the amount is outside the range {0..MAX_MONEY}.

pub fn from_u64_le_bytes(bytes: [u8; 8]) -> Result<Self, BalanceError>

Reads an Zatoshis from an unsigned 64-bit little-endian integer.

Returns an error if the amount is outside the range {0..MAX_MONEY}.

pub fn from_nonnegative_i64_le_bytes( bytes: [u8; 8], ) -> Result<Self, BalanceError>

Reads a Zatoshis from a signed integer represented as a two’s complement 64-bit little-endian value.

Returns an error if the amount is outside the range {0..MAX_MONEY}.

pub fn to_i64_le_bytes(self) -> [u8; 8]

Returns this Zatoshis encoded as a signed two’s complement 64-bit little-endian value.

pub fn is_zero(&self) -> bool

Returns whether or not this Zatoshis is the zero value.

pub fn is_positive(&self) -> bool

Returns whether or not this Zatoshis is positive.

pub fn div_with_remainder(&self, divisor: NonZeroU64) -> QuotRem<Zatoshis>

Divides this Zatoshis value by the given divisor and returns the quotient and remainder.