Crate zcash_protocol

A crate for Zcash protocol constants and value types.

zcash_protocol contains Rust structs, traits and functions that provide the network constants for the Zcash main and test networks, as well types for representing ZEC amounts and value balances.

Feature flags

test-dependencies — Exposes APIs that are useful for testing, such as proptest strategies.

local-consensus — Exposes support for working with a local consensus (e.g. regtest).

Modules

consensus

Consensus logic and parameters.

constants

Network-specific Zcash constants.

local_consensuslocal-consensus

memo

Structs for handling encrypted memos.

value

Enums

PoolType

A value pool in the Zcash protocol.

ShieldedProtocol

A Zcash shielded transfer protocol.