zcash_protocol
Module consensus

Consensus logic and parameters.

Modules

testingtest-dependencies

Structs

BlockHeight

A wrapper type representing blockchain heights.

MainNetwork

Marker struct for the production network.

TestNetwork

Marker struct for the test network.

Enums

BranchId

A globally-unique identifier for a set of consensus rules within the Zcash chain.

Network

The enumeration of known Zcash networks.

NetworkType

The enumeration of known Zcash network types.

NetworkUpgrade

An event that occurs at a specified height on the Zcash chain, at which point the consensus rules enforced by the network are altered.

Constants

H0

The height of the genesis block on a network.

MAIN_NETWORK

The production network.

TEST_NETWORK

The test network.

ZIP212_GRACE_PERIOD

The “grace period” defined in ZIP 212.

Traits

NetworkConstants

Constants associated with a given Zcash network.

Parameters

Zcash consensus parameters.